Replaces Bond.tid with u32. closes #7

+++ src/graph/bond.rs
@@ -3,11 +3,11 @@ use super::BondKind;
 #[derive(Debug, PartialEq, Clone)]
 pub struct Bond {
     pub kind: BondKind,
-    pub tid: usize,
+    pub tid: u32,
 }
 
 impl Bond {
-    pub fn new(kind: BondKind, tid: usize) -> Self {
+    pub fn new(kind: BondKind, tid: u32) -> Self {
         Self { kind, tid }
     }
